Rename Layout -> Common
